240613 서보원 DB생성, fk 연결, 더미데이터생성 완료

+++ HanJan/database/migrations/2024_06_12_153032_create_bags_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('bags', function (Blueprint $table) {
             $table->id('ba_id');
-            $table->integer('u_id')->unsigned();
-            $table->integer('p_id')->unsigned();
+            $table->bigInteger('u_id')->unsigned();
+            $table->bigInteger('p_id')->unsigned();
             $table->timestamp('created_at')->useCurrent();
             $table->softDeletes(); 
         });
